+++ inst/examples/ex-outliers.R
@@ -4,17 +4,12 @@ kommos <- remove_NA(kommos, margin = 1) # Remove cases with missing values
 coda <- as_composition(kommos, parts = 3:17, groups = 1)
 
 ## Detect outliers
-out <- outliers(coda, groups = NULL, method = "mcd")
+out <- outliers(coda, method = "mcd")
 
 plot(out, type = "dotchart")
 plot(out, type = "distance")
-plot(out, type = "qqplot")
 
-## Detect outliers by group
-## (use ceramic types for grouping)
-out <- outliers(coda, method = "mcd")
-
-plot(out, type = "dotchart", select = 1, robust = FALSE)
-plot(out, type = "dotchart", select = 2, robust = FALSE)
-plot(out, type = "dotchart", select = 3, robust = FALSE)
-plot(out, type = "dotchart", select = 4, robust = FALSE)
+## Detect outliers according to CJ
+ref <- extract(coda, "CJ")
+out <- outliers(coda, reference = ref, method = "mcd")
+plot(out, type = "dotchart")